+++ src/analysis/weightsToEdgeEvaluator.js
@@ -1,9 +1,9 @@
 // @flow
 
-import type {Edge} from "../../../core/graph";
-import type {WeightedTypes} from "../../../analysis/weights";
-import type {EdgeEvaluator} from "../../../analysis/pagerank";
-import {NodeTrie, EdgeTrie} from "../../../core/trie";
+import type {Edge} from "../core/graph";
+import type {WeightedTypes} from "./weights";
+import type {EdgeEvaluator} from "./pagerank";
+import {NodeTrie, EdgeTrie} from "../core/trie";
 
 export function weightsToEdgeEvaluator(weights: WeightedTypes): EdgeEvaluator {
   const nodeTrie = new NodeTrie();
